refactor service change detail entity package
