+++ modules/database/src/main/java/org/eclipse/xpanse/modules/database/servicechange/ServiceChangeDetailsRepository.java
@@ -4,7 +4,7 @@
  *
  */
 
-package org.eclipse.xpanse.modules.database.serviceconfiguration.update;
+package org.eclipse.xpanse.modules.database.servicechange;
 
 import jakarta.persistence.LockModeType;
 import jakarta.persistence.QueryHint;
