+++ modules/deployment/src/main/java/org/eclipse/xpanse/modules/deployment/ServiceActionManager.java
@@ -14,8 +14,8 @@ import java.util.UUID;
 import java.util.stream.Collectors;
 import lombok.extern.slf4j.Slf4j;
 import org.eclipse.xpanse.modules.database.service.ServiceDeploymentEntity;
-import org.eclipse.xpanse.modules.database.serviceconfiguration.update.ServiceChangeDetailsEntity;
-import org.eclipse.xpanse.modules.database.serviceconfiguration.update.ServiceChangeDetailsStorage;
+import org.eclipse.xpanse.modules.database.servicechange.ServiceChangeDetailsEntity;
+import org.eclipse.xpanse.modules.database.servicechange.ServiceChangeDetailsStorage;
 import org.eclipse.xpanse.modules.database.servicetemplate.ServiceTemplateEntity;
 import org.eclipse.xpanse.modules.database.servicetemplate.ServiceTemplateStorage;
 import org.eclipse.xpanse.modules.logging.CustomRequestIdGenerator;
