+++ modules/deployment/src/main/java/org/eclipse/xpanse/modules/deployment/ServiceChangeDetailsManager.java
@@ -17,9 +17,9 @@ import java.util.UUID;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.eclipse.xpanse.modules.database.service.ServiceDeploymentEntity;
-import org.eclipse.xpanse.modules.database.serviceconfiguration.update.ServiceChangeDetailsEntity;
-import org.eclipse.xpanse.modules.database.serviceconfiguration.update.ServiceChangeDetailsQueryModel;
-import org.eclipse.xpanse.modules.database.serviceconfiguration.update.ServiceChangeDetailsStorage;
+import org.eclipse.xpanse.modules.database.servicechange.ServiceChangeDetailsEntity;
+import org.eclipse.xpanse.modules.database.servicechange.ServiceChangeDetailsQueryModel;
+import org.eclipse.xpanse.modules.database.servicechange.ServiceChangeDetailsStorage;
 import org.eclipse.xpanse.modules.database.serviceorder.ServiceOrderEntity;
 import org.eclipse.xpanse.modules.database.serviceorder.ServiceOrderStorage;
 import org.eclipse.xpanse.modules.database.utils.EntityTransUtils;
